+++ gazon-frontend/src/main/java/com/gitlab/FrontendApplication.java
@@ -1,0 +1,14 @@
+package com.gitlab;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+
+@SpringBootApplication
+@EnableFeignClients
+public class FrontendApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(FrontendApplication.class, args);
+    }
+}